move game code inside Main.java to GameScene.java Now can swtich to game from mainMenu
